fix: tighten CORS credentials and monetary model typing
Agent-Logs-Url: https://github.com/ChibuikemJosh/SmartSync/sessions/d97e1d2a-4801-4f73-a659-ba7c4006a985

+++ models/squad.py
@@ -1,3 +1,5 @@
+from decimal import Decimal
+
 from pydantic import BaseModel
 
 
@@ -9,6 +11,6 @@ class VirtualAccountRequest(BaseModel):
 
 
 class PaymentLinkRequest(BaseModel):
-    amount: float
+    amount: Decimal
     currency: str = "NGN"
     description: str | None = None
